pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "happyfigure"
version = "0.1.0"
description = "AI-powered scientific figure generation pipeline"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.14"
authors = [
    {name = "HappyFigure Contributors"},
]
keywords = ["scientific-figures", "llm", "langgraph", "figure-generation", "svg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "langgraph>=0.2",
    "openai>=1.0",
    "python-dotenv",
    "Pillow>=10.0",
    "requests",
    "pydantic>=2.0",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pandas>=1.5",
]

[project.optional-dependencies]
google = ["google-genai>=1.0"]
anthropic = ["anthropic>=0.40"]
azure = ["azure-identity>=1.15"]
bedrock = ["boto3>=1.34"]
svg = [
    "cairosvg>=2.7",
    "lxml>=4.9",
]
svg-fallback = [
    # Requires system package on Linux: sudo apt install libcairo2-dev pkg-config
    "svglib>=1.5",
    "reportlab>=3.6",
]
examples = [
    "seaborn>=0.12",
    "scipy>=1.10",
    "python-dateutil>=2.8",
]
services = [
    "fastapi>=0.110",
    "uvicorn>=0.24",
    # torch: install separately with CUDA support BEFORE this extra:
    #   pip install torch --index-url https://download.pytorch.org/whl/cu126
    # Or on PyTorch>=2.6 from PyPI (ships with CUDA by default):
    #   pip install torch>=2.6
    "torch>=2.1",
    "opencv-python-headless>=4.9",
    "huggingface-hub>=0.20",
    "sam3>=0.1.2",
    "ben2 @ git+https://github.com/PramaLLC/BEN2.git@2c99a5da477b5523585bfa5c893888a6e818a8f6",
    # paddlepaddle-gpu 3.2.1 only available for Python <=3.11;
    # for Python 3.12+, use paddlepaddle (CPU) which works fine for OCR
    "paddlepaddle==3.2.1; python_version>='3.12'",
    "paddlepaddle-gpu==3.2.1; python_version<'3.12'",
    "paddleocr>=3.4; python_version<'3.13'",
    "paddlex>=3.4; python_version<'3.13'",
]
dev = ["ruff", "pytest"]
all = [
    "happyfigure[google]",
    "happyfigure[anthropic]",
    "happyfigure[azure]",
    "happyfigure[bedrock]",
    "happyfigure[svg]",
    "happyfigure[examples]",
]

[project.scripts]
happyfigure = "cli:main"

[project.urls]
Repository = "https://github.com/qingfengtommy/HappyFigure"

[tool.setuptools]
include-package-data = true
py-modules = ["cli"]

[tool.setuptools.packages.find]
include = ["llm*", "graphs*", "agents*", "tools*", "scripts*", "ui*",
           "services*",
           "configs*", "prompts*", "pipeline*"]

[tool.setuptools.package-data]
configs = ["*", "**/*"]
prompts = ["*", "**/*"]

[tool.ruff]
line-length = 120
target-version = "py310"
